Jul08
Issues with reading the SQLite file were resolved by creating a separate folder, copying the hawaii.sqlite file and app.py files into it, and copying the code into the app.py copy in this folder.

It appears that there were issues in the previous app.py folder with the Base.prepare function losing connection to the engine and failing to load the sqlite file as a class in Python. This copied file in the new folder appears to resolve the issue.